+++ internal/trillianclient/client.go
